requirements.txt
numpy>=1.24
PyYAML>=6.0
sentence-transformers>=2.7,<3.0
transformers>=4.41,<5.0
